requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
django-filter==23.5
gunicorn==21.2.0
